+++ src/main/java/com/nemonotfound/nemoscampfires/NemosCampfiresClient.java
@@ -12,6 +12,7 @@ public class NemosCampfiresClient implements ClientModInitializer {
         BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.ACACIA_CAMPFIRE, RenderType.cutout());
         BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.BIRCH_CAMPFIRE, RenderType.cutout());
         BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.CHERRY_CAMPFIRE, RenderType.cutout());
+        BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.PALE_OAK_CAMPFIRE, RenderType.cutout());
         BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.CRIMSON_CAMPFIRE, RenderType.cutout());
         BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.DARK_OAK_CAMPFIRE, RenderType.cutout());
         BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.JUNGLE_CAMPFIRE, RenderType.cutout());
@@ -21,6 +22,7 @@ public class NemosCampfiresClient implements ClientModInitializer {
         BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.ACACIA_SOUL_CAMPFIRE, RenderType.cutout());
         BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.BIRCH_SOUL_CAMPFIRE, RenderType.cutout());
         BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.CHERRY_SOUL_CAMPFIRE, RenderType.cutout());
+        BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.PALE_OAK_SOUL_CAMPFIRE, RenderType.cutout());
         BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.CRIMSON_SOUL_CAMPFIRE, RenderType.cutout());
         BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.DARK_OAK_SOUL_CAMPFIRE, RenderType.cutout());
         BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.JUNGLE_SOUL_CAMPFIRE, RenderType.cutout());