+++ src/main/java/com/nemonotfound/nemoscampfires/NemosCampfiresDataGenerator.java
@@ -5,6 +5,7 @@ import com.nemonotfound.nemoscampfires.datagen.LootTableProvider;
 import com.nemonotfound.nemoscampfires.datagen.ModRecipeProvider;
 import com.nemonotfound.nemoscampfires.datagen.langdatagen.EnglishLanguageProvider;
 import com.nemonotfound.nemoscampfires.datagen.langdatagen.GermanLanguageProvider;
+import com.nemonotfound.nemoscampfires.datagen.langdatagen.RussianLanguageProvider;
 import com.nemonotfound.nemoscampfires.datagen.modelgen.ModelProvider;
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
@@ -17,6 +18,7 @@ public class NemosCampfiresDataGenerator implements DataGeneratorEntrypoint {
 
 		pack.addProvider(EnglishLanguageProvider::new);
 		pack.addProvider(GermanLanguageProvider::new);
+		pack.addProvider(RussianLanguageProvider::new);
 		pack.addProvider(BlockTagProvider::new);
 		pack.addProvider(ModelProvider::new);
 		pack.addProvider(LootTableProvider::new);