+++ Changelog.md
@@ -1,7 +1,11 @@
-# Changelog - v1.2.1
+# Changelog - v1.3
 
-## Changes
-- Custom campfire now uses 100% vanilla code
+## Additions
+- Added pale oak campfire
+- Added russian translation
+
+# Changes
+- Slightly adapted almost every campfire item texture
 
 ### Downloads
 [CurseForge](https://www.curseforge.com/minecraft/mc-mods/nemos-campfires) <br>